Fix all compiler warnings (0 warnings on x86_64 and ARM64)
* Change encode/decode `len` params from int32_t to uint32_t across
  SetBool/UInt64/PingPong/Bulk, matching the tt_REQUEST_ENCODE/DECODE
  etc. typedefs they're cast to instead of casting around them
* Drop SetBool's always-false upper-bound check on message_length
  (uint16_t can't exceed tt_MAX_STRING_LENGTH by type anyway)
* Cast timeout comparisons in tt_Node_poll/tt_receive to uint64_t
  (already invariant-safe, just needed the sign to match)
* Verified zero warnings on x86_64 and both ARM64 Raspberry Pis

+++ src/hal_linux.c
@@ -138,7 +138,7 @@ int32_t tt_receive(struct tt_Node* node, void* buf, size_t len, uint32_t* ip, ui
     struct sockaddr_in addr;
     socklen_t addr_len = sizeof(struct sockaddr_in);
 
-    if (timeout >= 0 && node->hal.receive_timeout != timeout) {
+    if (timeout >= 0 && node->hal.receive_timeout != (uint64_t)timeout) {
         struct timeval tval;
         tval.tv_sec = timeout / SEC_NS;
         tval.tv_usec = (timeout % SEC_NS) / US_NS; // nano to micro
